Split the code that was written together in `book.js` into separate files. 🍕
